+++ agent/api/task/task_windows.go
@@ -16,13 +16,15 @@
 package task
 
 import (
-	"errors"
 	"runtime"
 	"time"
 
+	"github.com/aws/amazon-ecs-agent/agent/ecscni"
 	"github.com/aws/amazon-ecs-agent/agent/utils"
+	"github.com/containernetworking/cni/libcni"
 
 	apicontainerstatus "github.com/aws/amazon-ecs-agent/agent/api/container/status"
+	apieni "github.com/aws/amazon-ecs-agent/agent/api/eni"
 	"github.com/aws/amazon-ecs-agent/agent/config"
 	"github.com/aws/amazon-ecs-agent/agent/credentials"
 	"github.com/aws/amazon-ecs-agent/agent/taskresource"
@@ -33,6 +35,7 @@ import (
 	taskresourcevolume "github.com/aws/amazon-ecs-agent/agent/taskresource/volume"
 	"github.com/cihub/seelog"
 	dockercontainer "github.com/docker/docker/api/types/container"
+	"github.com/pkg/errors"
 )
 
 const (
@@ -245,4 +248,43 @@ func (task *Task) addFSxWindowsFileServerResource(
 	task.updateContainerVolumeDependency(vol.Name)
 
 	return nil
+}
+
+// BuildCNIConfig builds a list of CNI network configurations for the task.
+// The first configuration is for vpc-shared-eni plugin to setup the task eni in task namespace
+func (task *Task) BuildCNIConfig(includeIPAMConfig bool, cniConfig *ecscni.Config) (*ecscni.Config, error) {
+	if !task.IsNetworkModeAWSVPC() {
+		return nil, errors.New("task config: task network mode is not awsvpc")
+	}
+
+	var netconf *libcni.NetworkConfig
+	var err error
+
+	// Build a CNI network configuration for each ENI.
+	for _, eni := range task.ENIs {
+		switch eni.InterfaceAssociationProtocol {
+		// If the association protocol is set to "default" or unset (to preserve backwards
+		// compatibility), consider it a "standard" ENI attachment.
+		case "", apieni.DefaultInterfaceAssociationProtocol:
+			cniConfig.ID = eni.MacAddress
+			netconf, err = ecscni.NewBridgeNetworkConfigForTaskNSSetup(eni, cniConfig)
+		default:
+			err = errors.Errorf("task config: unknown interface association type: %s",
+				eni.InterfaceAssociationProtocol)
+		}
+
+		if err != nil {
+			return nil, err
+		}
+
+		// IfName can be an empty string for Windows CNI plugins as it is not used for naming HNS endpoints
+		cniConfig.NetworkConfigs = append(cniConfig.NetworkConfigs, &ecscni.NetworkConfig{
+			IfName:           ecscni.TaskENIBridgeNetworkPrefix,
+			CNINetworkConfig: netconf,
+		})
+	}
+
+	// TODO: Build configuration for second invocation of plugin to setup ecs-bridge
+
+	return cniConfig, nil
 }
