+++ agent/app/agent_capability_test.go
@@ -24,13 +24,11 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/aws/amazon-ecs-agent/agent/ecs_client/model/ecs"
-
 	app_mocks "github.com/aws/amazon-ecs-agent/agent/app/mocks"
 	"github.com/aws/amazon-ecs-agent/agent/config"
 	"github.com/aws/amazon-ecs-agent/agent/dockerclient"
 	mock_dockerapi "github.com/aws/amazon-ecs-agent/agent/dockerclient/dockerapi/mocks"
-	"github.com/aws/amazon-ecs-agent/agent/ecscni"
+	"github.com/aws/amazon-ecs-agent/agent/ecs_client/model/ecs"
 	mock_ecscni "github.com/aws/amazon-ecs-agent/agent/ecscni/mocks"
 	mock_pause "github.com/aws/amazon-ecs-agent/agent/eni/pause/mocks"
 	mock_mobypkgwrapper "github.com/aws/amazon-ecs-agent/agent/utils/mobypkgwrapper/mocks"
@@ -103,7 +101,9 @@ func TestCapabilities(t *testing.T) {
 			dockerclient.Version_1_18,
 			dockerclient.Version_1_19,
 		}),
-		cniClient.EXPECT().Version(ecscni.ECSENIPluginName).Return("v1", nil),
+		// CNI plugins are platform dependent.
+		// Therefore, for any version query for any plugin return an appropriate version
+		cniClient.EXPECT().Version(gomock.Any()).Return("v1", nil),
 		mockMobyPlugins.EXPECT().Scan().AnyTimes().Return([]string{}, nil),
 		client.EXPECT().ListPluginsWithFilters(gomock.Any(), gomock.Any(), gomock.Any(),
 			gomock.Any()).AnyTimes().Return([]string{}, nil),
@@ -222,7 +222,8 @@ func getCapabilitiesWithConfig(cfg *config.Config, t *testing.T) []*ecs.Attribut
 	mockCNIClient := mock_ecscni.NewMockCNIClient(ctrl)
 
 	mockPauseLoader.EXPECT().IsLoaded(gomock.Any()).Return(false, nil).AnyTimes()
-	mockCNIClient.EXPECT().Version(ecscni.ECSENIPluginName).Return("v1", nil).AnyTimes()
+	// CNI plugins are platform dependent. Therefore return version for any plugin query.
+	mockCNIClient.EXPECT().Version(gomock.Any()).Return("v1", nil).AnyTimes()
 	gomock.InOrder(
 		client.EXPECT().SupportedVersions().Return([]dockerclient.DockerVersion{
 			dockerclient.Version_1_17,
@@ -564,7 +565,9 @@ func TestCapabilitiesExecutionRoleAWSLogs(t *testing.T) {
 		dockerclient.Version_1_17,
 	})
 	client.EXPECT().KnownVersions().Return(nil)
-	cniClient.EXPECT().Version(ecscni.ECSENIPluginName).Return("v1", errors.New("some error happened"))
+	// CNI plugins are platform dependent.
+	// Therefore, for any version query for any plugin return an error
+	cniClient.EXPECT().Version(gomock.Any()).Return("v1", errors.New("some error happened"))
 	mockMobyPlugins.EXPECT().Scan().AnyTimes().Return([]string{}, nil)
 	client.EXPECT().ListPluginsWithFilters(gomock.Any(), gomock.Any(), gomock.Any(),
 		gomock.Any()).AnyTimes().Return([]string{}, nil)
