+++ agent/app/agent_capability_windows_test.go
@@ -72,7 +72,7 @@ func TestVolumeDriverCapabilitiesWindows(t *testing.T) {
 			dockerclient.Version_1_18,
 			dockerclient.Version_1_19,
 		}),
-		cniClient.EXPECT().Version(ecscni.ECSENIPluginName).Return("v1", nil),
+		cniClient.EXPECT().Version(ecscni.ECSVPCSharedENIPluginExecutable).Return("v1", nil),
 	)
 
 	expectedCapabilityNames := []string{
@@ -160,7 +160,7 @@ func TestSupportedCapabilitiesWindows(t *testing.T) {
 			dockerclient.Version_1_18,
 			dockerclient.Version_1_19,
 		}),
-		cniClient.EXPECT().Version(ecscni.ECSENIPluginName).Return("v1", nil),
+		cniClient.EXPECT().Version(ecscni.ECSVPCSharedENIPluginExecutable).Return("v1", nil),
 	)
 
 	expectedCapabilityNames := []string{
