+++ agent/config/config_windows.go
@@ -55,6 +55,8 @@ const (
 	minimumContainerCreateTimeout = 1 * time.Minute
 	// default image pull inactivity time is extra time needed on container extraction
 	defaultImagePullInactivityTimeout = 3 * time.Minute
+	// default directory name of CNI Plugins
+	defaultCNIPluginDirName = "cni"
 )
 
 // DefaultConfig returns the default configuration for Windows
@@ -62,6 +64,10 @@ func DefaultConfig() Config {
 	programData := utils.DefaultIfBlank(os.Getenv("ProgramData"), `C:\ProgramData`)
 	ecsRoot := filepath.Join(programData, "Amazon", "ECS")
 	dataDir := filepath.Join(ecsRoot, "data")
+
+	programFiles := utils.DefaultIfBlank(os.Getenv("ProgramFiles"), `C:\Program Files`)
+	ecsBinaryDir := filepath.Join(programFiles, "Amazon", "ECS")
+
 	platformVariables := PlatformVariables{
 		CPUUnbounded:    BooleanDefaultFalse{Value: ExplicitlyDisabled},
 		MemoryUnbounded: BooleanDefaultFalse{Value: ExplicitlyDisabled},
@@ -115,6 +121,7 @@ func DefaultConfig() Config {
 		FSxWindowsFileServerCapable:         true,
 		PauseContainerImageName:             DefaultPauseContainerImageName,
 		PauseContainerTag:                   DefaultPauseContainerTag,
+		CNIPluginsPath:                      filepath.Join(ecsBinaryDir, defaultCNIPluginDirName),
 	}
 }
