+++ agent/ecscni/netconfig_windows.go
@@ -1,0 +1,84 @@
+// +build windows
+
+// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package ecscni
+
+import (
+	"net"
+
+	"github.com/containernetworking/cni/pkg/types"
+
+	"github.com/aws/amazon-ecs-agent/agent/api/eni"
+	"github.com/containernetworking/cni/libcni"
+	"github.com/pkg/errors"
+)
+
+// NewBridgeNetworkConfigForTaskNSSetup is used to create the bridge configuration for task namespace setup
+func NewBridgeNetworkConfigForTaskNSSetup(eni *eni.ENI, cfg *Config) (*libcni.NetworkConfig, error) {
+	gateway := eni.GetSubnetGatewayIPv4Address()
+	dns := types.DNS{
+		Nameservers: eni.DomainNameServers,
+	}
+
+	if len(eni.DomainNameServers) == 0 {
+		if gateway == "" {
+			return nil, errors.New("cannot create bridge network config due to invalid gateway")
+		}
+		constructedDNS, err := constructDNSFromVPCGateway(gateway)
+		if err != nil {
+			return nil, errors.Wrapf(err, "cannot create bridge network config")
+		}
+		dns.Nameservers = constructedDNS
+	}
+
+	eniConf := BridgeForTaskENIConfig{
+		Type:             ECSVPCSharedENIPluginName,
+		ENIIPAddress:     eni.GetPrimaryIPv4AddressWithPrefixLength(),
+		ENIMACAddress:    eni.MacAddress,
+		GatewayIPAddress: gateway,
+		IPAddress:        eni.GetPrimaryIPv4AddressWithPrefixLength(),
+		TaskENIConfig: TaskENIConfig{
+			PauseContainer: true,
+		},
+		DNS: dns,
+	}
+
+	networkConfig, err := newNetworkConfig(eniConf, ECSVPCSharedENIPluginExecutable, cfg.MinSupportedCNIVersion)
+	if err != nil {
+		return nil, errors.Wrap(err, "cni config: failed to create bridge plugin configuration for setting up task network namespace")
+	}
+
+	// Network.Name is used to create the name of the task bridge network.
+	// The bridge name generated by the plugin would be of format taskbr<ENI MAC Address>
+	networkConfig.Network.Name = TaskENIBridgeNetworkPrefix
+
+	return networkConfig, nil
+}
+
+// constructDNSFromVPCGateway is used to construct DNS server from subnet gateway address
+func constructDNSFromVPCGateway(subnetGateway string) ([]string, error) {
+	// The DNS server maps to a reserved IP address at the base of the VPC IPv4 network rage plus 2
+	// https://docs.aws.amazon.com/vpc/latest/userguide/VPC_DHCP_Options.html#AmazonDNS
+	// Therefore, we construct the same by replacing the last token of the subnet gateway to 2
+	subnetIPv4 := net.ParseIP(subnetGateway)
+	if subnetIPv4 == nil {
+		return nil, errors.New("failed to construct dns from gateway address")
+	}
+	mask := net.CIDRMask(24, 32)
+	maskedIPv4 := subnetIPv4.Mask(mask).To4()
+	maskedIPv4[3] = 2
+
+	return []string{maskedIPv4.String()}, nil
+}
