+++ agent/engine/docker_task_engine_test.go
@@ -46,7 +46,6 @@ import (
 	"github.com/aws/amazon-ecs-agent/agent/dockerclient"
 	"github.com/aws/amazon-ecs-agent/agent/dockerclient/dockerapi"
 	mock_dockerapi "github.com/aws/amazon-ecs-agent/agent/dockerclient/dockerapi/mocks"
-	"github.com/aws/amazon-ecs-agent/agent/ecscni"
 	mock_ecscni "github.com/aws/amazon-ecs-agent/agent/ecscni/mocks"
 	"github.com/aws/amazon-ecs-agent/agent/engine/dockerstate"
 	"github.com/aws/amazon-ecs-agent/agent/engine/execcmd"
@@ -104,6 +103,7 @@ const (
 	networkModeBridge           = "bridge"
 	networkModeAWSVPC           = "awsvpc"
 	testTaskARN                 = "arn:aws:ecs:region:account-id:task/task-id"
+	containerNetworkMode        = "none"
 )
 
 var (
@@ -334,140 +334,6 @@ func TestBatchContainerHappyPath(t *testing.T) {
 			}
 		})
 	}
-}
-
-// TestTaskWithSteadyStateResourcesProvisioned tests container and task transitions
-// when the steady state for the pause container is set to RESOURCES_PROVISIONED and
-// the steady state for the normal container is set to RUNNING
-func TestTaskWithSteadyStateResourcesProvisioned(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.TODO())
-	defer cancel()
-	ctrl, client, mockTime, taskEngine, _, imageManager, _ := mocks(t, ctx, &defaultConfig)
-	defer ctrl.Finish()
-
-	mockCNIClient := mock_ecscni.NewMockCNIClient(ctrl)
-	taskEngine.(*DockerTaskEngine).cniClient = mockCNIClient
-	// sleep5 contains a single 'sleep' container, with DesiredStatus == RUNNING
-	sleepTask := testdata.LoadTask("sleep5")
-	sleepContainer := sleepTask.Containers[0]
-	sleepContainer.TransitionDependenciesMap = make(map[apicontainerstatus.ContainerStatus]apicontainer.TransitionDependencySet)
-	sleepContainer.BuildContainerDependency("pause", apicontainerstatus.ContainerResourcesProvisioned, apicontainerstatus.ContainerPulled)
-	// Add a second container with DesiredStatus == RESOURCES_PROVISIONED and
-	// steadyState == RESOURCES_PROVISIONED
-	pauseContainer := apicontainer.NewContainerWithSteadyState(apicontainerstatus.ContainerResourcesProvisioned)
-	pauseContainer.Name = "pause"
-	pauseContainer.Image = "pause"
-	pauseContainer.CPU = 10
-	pauseContainer.Memory = 10
-	pauseContainer.Essential = true
-	pauseContainer.Type = apicontainer.ContainerCNIPause
-	sleepTask.Containers = append(sleepTask.Containers, pauseContainer)
-	eventStream := make(chan dockerapi.DockerContainerChangeEvent)
-	// containerEventsWG is used to force the test to wait until the container created and started
-	// events are processed
-	containerEventsWG := sync.WaitGroup{}
-
-	client.EXPECT().ContainerEvents(gomock.Any()).Return(eventStream, nil)
-	// We cannot rely on the order of pulls between images as they can still be downloaded in
-	// parallel. The dependency graph enforcement comes into effect for CREATED transitions.
-	// Hence, do not enforce the order of invocation of these calls
-	imageManager.EXPECT().AddAllImageStates(gomock.Any()).AnyTimes()
-	client.EXPECT().PullImage(gomock.Any(), sleepContainer.Image, nil, gomock.Any()).Return(dockerapi.DockerContainerMetadata{})
-	imageManager.EXPECT().RecordContainerReference(sleepContainer).Return(nil)
-	imageManager.EXPECT().GetImageStateFromImageName(sleepContainer.Image).Return(nil, false)
-
-	gomock.InOrder(
-		// Ensure that the pause container is created first
-		client.EXPECT().APIVersion().Return(defaultDockerClientAPIVersion, nil),
-		client.EXPECT().CreateContainer(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Do(
-			func(ctx interface{}, config *dockercontainer.Config, hostConfig *dockercontainer.HostConfig, containerName string, z time.Duration) {
-				sleepTask.AddTaskENI(mockENI)
-				sleepTask.SetAppMesh(&appmesh.AppMesh{
-					IgnoredUID:       ignoredUID,
-					ProxyIngressPort: proxyIngressPort,
-					ProxyEgressPort:  proxyEgressPort,
-					AppPorts: []string{
-						appPort,
-					},
-					EgressIgnoredIPs: []string{
-						egressIgnoredIP,
-					},
-				})
-				assert.Equal(t, "none", string(hostConfig.NetworkMode))
-				assert.True(t, strings.Contains(containerName, pauseContainer.Name))
-				containerEventsWG.Add(1)
-				go func() {
-					eventStream <- createDockerEvent(apicontainerstatus.ContainerCreated)
-					containerEventsWG.Done()
-				}()
-			}).Return(dockerapi.DockerContainerMetadata{DockerID: containerID + ":" + pauseContainer.Name}),
-		// Ensure that the pause container is started after it's created
-		client.EXPECT().StartContainer(gomock.Any(), containerID+":"+pauseContainer.Name, defaultConfig.ContainerStartTimeout).Do(
-			func(ctx interface{}, id string, timeout time.Duration) {
-				containerEventsWG.Add(1)
-				go func() {
-					eventStream <- createDockerEvent(apicontainerstatus.ContainerRunning)
-					containerEventsWG.Done()
-				}()
-			}).Return(dockerapi.DockerContainerMetadata{DockerID: containerID + ":" + pauseContainer.Name}),
-		client.EXPECT().InspectContainer(gomock.Any(), gomock.Any(), gomock.Any()).Return(&types.ContainerJSON{
-			ContainerJSONBase: &types.ContainerJSONBase{
-				ID:    containerID,
-				State: &types.ContainerState{Pid: 23},
-			},
-		}, nil),
-		// Then setting up the pause container network namespace
-		mockCNIClient.EXPECT().SetupNS(gomock.Any(), gomock.Any(), gomock.Any()).Return(nsResult, nil),
-
-		// Once the pause container is started, sleep container will be created
-		client.EXPECT().APIVersion().Return(defaultDockerClientAPIVersion, nil),
-		client.EXPECT().CreateContainer(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Do(
-			func(ctx interface{}, config *dockercontainer.Config, hostConfig *dockercontainer.HostConfig, containerName string, z time.Duration) {
-				assert.True(t, strings.Contains(containerName, sleepContainer.Name))
-				assert.Equal(t, "container:"+containerID+":"+pauseContainer.Name, string(hostConfig.NetworkMode))
-				containerEventsWG.Add(1)
-				go func() {
-					eventStream <- createDockerEvent(apicontainerstatus.ContainerCreated)
-					containerEventsWG.Done()
-				}()
-			}).Return(dockerapi.DockerContainerMetadata{DockerID: containerID + ":" + sleepContainer.Name}),
-		// Next, the sleep container is started
-		client.EXPECT().StartContainer(gomock.Any(), containerID+":"+sleepContainer.Name, defaultConfig.ContainerStartTimeout).Do(
-			func(ctx interface{}, id string, timeout time.Duration) {
-				containerEventsWG.Add(1)
-				go func() {
-					eventStream <- createDockerEvent(apicontainerstatus.ContainerRunning)
-					containerEventsWG.Done()
-				}()
-			}).Return(dockerapi.DockerContainerMetadata{DockerID: containerID + ":" + sleepContainer.Name}),
-	)
-
-	addTaskToEngine(t, ctx, taskEngine, sleepTask, mockTime, &containerEventsWG)
-	taskARNByIP, ok := taskEngine.(*DockerTaskEngine).state.GetTaskByIPAddress(taskIP)
-	assert.True(t, ok)
-	assert.Equal(t, sleepTask.Arn, taskARNByIP)
-	cleanup := make(chan time.Time, 1)
-	mockTime.EXPECT().After(gomock.Any()).Return(cleanup).AnyTimes()
-	client.EXPECT().InspectContainer(gomock.Any(), gomock.Any(), gomock.Any()).Return(
-		&types.ContainerJSON{
-			ContainerJSONBase: &types.ContainerJSONBase{
-				ID:    containerID,
-				State: &types.ContainerState{Pid: 23},
-			},
-		}, nil)
-	mockCNIClient.EXPECT().CleanupNS(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil)
-	client.EXPECT().StopContainer(gomock.Any(), containerID+":"+pauseContainer.Name, gomock.Any()).MinTimes(1)
-	mockCNIClient.EXPECT().ReleaseIPResource(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil).MaxTimes(1)
-
-	// Simulate a container stop event from docker
-	eventStream <- dockerapi.DockerContainerChangeEvent{
-		Status: apicontainerstatus.ContainerStopped,
-		DockerContainerMetadata: dockerapi.DockerContainerMetadata{
-			DockerID: containerID + ":" + sleepContainer.Name,
-			ExitCode: aws.Int(exitCode),
-		},
-	}
-	waitForStopEvents(t, taskEngine.StateChangeEvents(), true, false)
 }
 
 // TestRemoveEvents tests if the task engine can handle task events while the task is being
@@ -1127,185 +993,7 @@ func TestGetTaskByArn(t *testing.T) {
 	assert.False(t, found, "Task with invalid arn found in the task engine")
 }
 
-func TestPauseContainerHappyPath(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.TODO())
-	defer cancel()
-	ctrl, dockerClient, mockTime, taskEngine, _, imageManager, _ := mocks(t, ctx, &defaultConfig)
-	defer ctrl.Finish()
-
-	cniClient := mock_ecscni.NewMockCNIClient(ctrl)
-	taskEngine.(*DockerTaskEngine).cniClient = cniClient
-	taskEngine.(*DockerTaskEngine).taskSteadyStatePollInterval = taskSteadyStatePollInterval
-	eventStream := make(chan dockerapi.DockerContainerChangeEvent)
-	sleepTask := testdata.LoadTask("sleep5TwoContainers")
-	sleepContainer1 := sleepTask.Containers[0]
-	sleepContainer1.TransitionDependenciesMap = make(map[apicontainerstatus.ContainerStatus]apicontainer.TransitionDependencySet)
-	sleepContainer2 := sleepTask.Containers[1]
-	sleepContainer2.TransitionDependenciesMap = make(map[apicontainerstatus.ContainerStatus]apicontainer.TransitionDependencySet)
-
-	// Add eni information to the task so the task can add dependency of pause container
-	sleepTask.AddTaskENI(mockENI)
-
-	sleepTask.SetAppMesh(&appmesh.AppMesh{
-		IgnoredUID:       ignoredUID,
-		ProxyIngressPort: proxyIngressPort,
-		ProxyEgressPort:  proxyEgressPort,
-		AppPorts: []string{
-			appPort,
-		},
-		EgressIgnoredIPs: []string{
-			egressIgnoredIP,
-		},
-	})
-
-	dockerClient.EXPECT().ContainerEvents(gomock.Any()).Return(eventStream, nil)
-
-	sleepContainerID1 := containerID + "1"
-	sleepContainerID2 := containerID + "2"
-	pauseContainerID := "pauseContainerID"
-	// Pause container will be launched first
-	gomock.InOrder(
-		dockerClient.EXPECT().APIVersion().Return(defaultDockerClientAPIVersion, nil),
-		dockerClient.EXPECT().CreateContainer(
-			gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Do(
-			func(ctx interface{}, config *dockercontainer.Config, x, y, z interface{}) {
-				name, ok := config.Labels[labelPrefix+"container-name"]
-				assert.True(t, ok)
-				assert.Equal(t, apitask.NetworkPauseContainerName, name)
-			}).Return(dockerapi.DockerContainerMetadata{DockerID: "pauseContainerID"}),
-		dockerClient.EXPECT().StartContainer(gomock.Any(), pauseContainerID, defaultConfig.ContainerStartTimeout).Return(
-			dockerapi.DockerContainerMetadata{DockerID: "pauseContainerID"}),
-		dockerClient.EXPECT().InspectContainer(gomock.Any(), gomock.Any(), gomock.Any()).Return(
-			&types.ContainerJSON{
-				ContainerJSONBase: &types.ContainerJSONBase{
-					ID:    pauseContainerID,
-					State: &types.ContainerState{Pid: containerPid},
-				},
-			}, nil),
-		cniClient.EXPECT().SetupNS(gomock.Any(), gomock.Any(), gomock.Any()).Return(nsResult, nil),
-	)
-
-	// For the other container
-	imageManager.EXPECT().AddAllImageStates(gomock.Any()).AnyTimes()
-	dockerClient.EXPECT().PullImage(gomock.Any(), gomock.Any(), nil, gomock.Any()).Return(dockerapi.DockerContainerMetadata{}).Times(2)
-	imageManager.EXPECT().RecordContainerReference(gomock.Any()).Return(nil).Times(2)
-	imageManager.EXPECT().GetImageStateFromImageName(gomock.Any()).Return(nil, false).Times(2)
-	dockerClient.EXPECT().APIVersion().Return(defaultDockerClientAPIVersion, nil).Times(2)
-
-	dockerClient.EXPECT().CreateContainer(gomock.Any(), gomock.Any(), gomock.Any(),
-		gomock.Any(), gomock.Any()).Return(dockerapi.DockerContainerMetadata{DockerID: sleepContainerID1})
-	dockerClient.EXPECT().CreateContainer(gomock.Any(), gomock.Any(), gomock.Any(),
-		gomock.Any(), gomock.Any()).Return(dockerapi.DockerContainerMetadata{DockerID: sleepContainerID2})
-
-	dockerClient.EXPECT().StartContainer(gomock.Any(), sleepContainerID1, defaultConfig.ContainerStartTimeout).Return(
-		dockerapi.DockerContainerMetadata{DockerID: sleepContainerID1})
-	dockerClient.EXPECT().StartContainer(gomock.Any(), sleepContainerID2, defaultConfig.ContainerStartTimeout).Return(
-		dockerapi.DockerContainerMetadata{DockerID: sleepContainerID2})
-
-	cleanup := make(chan time.Time)
-	defer close(cleanup)
-	mockTime.EXPECT().Now().Return(time.Now()).MinTimes(1)
-	dockerClient.EXPECT().DescribeContainer(gomock.Any(), sleepContainerID1).AnyTimes()
-	dockerClient.EXPECT().DescribeContainer(gomock.Any(), sleepContainerID2).AnyTimes()
-	dockerClient.EXPECT().DescribeContainer(gomock.Any(), pauseContainerID).AnyTimes()
-
-	err := taskEngine.Init(ctx)
-	assert.NoError(t, err)
-
-	taskEngine.AddTask(sleepTask)
-	stateChangeEvents := taskEngine.StateChangeEvents()
-	verifyTaskIsRunning(stateChangeEvents, sleepTask)
-
-	var wg sync.WaitGroup
-	wg.Add(1)
-	mockTime.EXPECT().After(gomock.Any()).Return(cleanup).MinTimes(1)
-
-	gomock.InOrder(
-		dockerClient.EXPECT().StopContainer(gomock.Any(), sleepContainerID2, gomock.Any()).Return(
-			dockerapi.DockerContainerMetadata{DockerID: sleepContainerID2}),
-
-		dockerClient.EXPECT().InspectContainer(gomock.Any(), pauseContainerID, gomock.Any()).Return(&types.ContainerJSON{
-			ContainerJSONBase: &types.ContainerJSONBase{
-				ID:    pauseContainerID,
-				State: &types.ContainerState{Pid: containerPid},
-			},
-		}, nil),
-		cniClient.EXPECT().CleanupNS(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil),
-
-		dockerClient.EXPECT().StopContainer(gomock.Any(), pauseContainerID, gomock.Any()).Return(
-			dockerapi.DockerContainerMetadata{DockerID: pauseContainerID}),
-
-		cniClient.EXPECT().ReleaseIPResource(gomock.Any(), gomock.Any(), gomock.Any()).Do(
-			func(ctx context.Context, cfg *ecscni.Config, timeout time.Duration) {
-				wg.Done()
-			}).Return(nil),
-	)
-
-	dockerClient.EXPECT().RemoveContainer(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil).Times(3)
-	imageManager.EXPECT().RemoveContainerReferenceFromImageState(gomock.Any()).Return(nil).Times(2)
-
-	// Simulate a container stop event from docker
-	eventStream <- dockerapi.DockerContainerChangeEvent{
-		Status: apicontainerstatus.ContainerStopped,
-		DockerContainerMetadata: dockerapi.DockerContainerMetadata{
-			DockerID: sleepContainerID1,
-			ExitCode: aws.Int(exitCode),
-		},
-	}
-
-	verifyTaskIsStopped(stateChangeEvents, sleepTask)
-	sleepTask.SetSentStatus(apitaskstatus.TaskStopped)
-	cleanup <- time.Now()
-	for {
-		tasks, _ := taskEngine.(*DockerTaskEngine).ListTasks()
-		if len(tasks) == 0 {
-			break
-		}
-		t.Logf("Found %d tasks in the engine; first task arn: %s", len(tasks), tasks[0].Arn)
-		fmt.Printf("Found %d tasks in the engine; first task arn: %s\n", len(tasks), tasks[0].Arn)
-		time.Sleep(5 * time.Millisecond)
-	}
-	wg.Wait()
-}
-
-func TestBuildCNIConfigFromTaskContainer(t *testing.T) {
-	config := defaultConfig
-	ctx, cancel := context.WithCancel(context.TODO())
-	defer cancel()
-	ctrl, _, _, taskEngine, _, _, _ := mocks(t, ctx, &config)
-	defer ctrl.Finish()
-
-	testTask := testdata.LoadTask("sleep5")
-	testTask.AddTaskENI(mockENI)
-	testTask.SetAppMesh(&appmesh.AppMesh{
-		IgnoredUID:       ignoredUID,
-		ProxyIngressPort: proxyIngressPort,
-		ProxyEgressPort:  proxyEgressPort,
-		AppPorts: []string{
-			appPort,
-		},
-		EgressIgnoredIPs: []string{
-			egressIgnoredIP,
-		},
-	})
-	containerInspectOutput := &types.ContainerJSON{
-		ContainerJSONBase: &types.ContainerJSONBase{
-			ID:    containerID,
-			State: &types.ContainerState{Pid: containerPid},
-		},
-	}
-
-	cniConfig, err := taskEngine.(*DockerTaskEngine).buildCNIConfigFromTaskContainer(testTask, containerInspectOutput, true)
-	assert.NoError(t, err)
-	assert.Equal(t, containerID, cniConfig.ContainerID)
-	assert.Equal(t, strconv.Itoa(containerPid), cniConfig.ContainerPID)
-	assert.Equal(t, mac, cniConfig.ID, "ID should be set to the mac of eni")
-	// We expect 3 NetworkConfig objects in the cni Config wrapper object:
-	// ENI, Bridge and Appmesh
-	require.Len(t, cniConfig.NetworkConfigs, 3)
-}
-
-func TestProvisionContainerResources(t *testing.T) {
+func TestProvisionContainerResourcesSetPausePIDInVolumeResources(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.TODO())
 	defer cancel()
 	ctrl, dockerClient, _, taskEngine, _, _, _ := mocks(t, ctx, &defaultConfig)
@@ -1340,6 +1028,9 @@ func TestProvisionContainerResources(t *testing.T) {
 			ContainerJSONBase: &types.ContainerJSONBase{
 				ID:    containerID,
 				State: &types.ContainerState{Pid: containerPid},
+				HostConfig: &dockercontainer.HostConfig{
+					NetworkMode: containerNetworkMode,
+				},
 			},
 		}, nil),
 		mockCNIClient.EXPECT().SetupNS(gomock.Any(), gomock.Any(), gomock.Any()).Return(nsResult, nil),
@@ -1421,6 +1112,9 @@ func TestStopPauseContainerCleanupCalled(t *testing.T) {
 			ContainerJSONBase: &types.ContainerJSONBase{
 				ID:    containerID,
 				State: &types.ContainerState{Pid: containerPid},
+				HostConfig: &dockercontainer.HostConfig{
+					NetworkMode: containerNetworkMode,
+				},
 			},
 		}, nil),
 		mockCNIClient.EXPECT().CleanupNS(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil),
@@ -1472,6 +1166,9 @@ func TestStopPauseContainerCleanupDelay(t *testing.T) {
 			ContainerJSONBase: &types.ContainerJSONBase{
 				ID:    containerID,
 				State: &types.ContainerState{Pid: containerPid},
+				HostConfig: &dockercontainer.HostConfig{
+					NetworkMode: containerNetworkMode,
+				},
 			},
 		}, nil),
 		mockCNIClient.EXPECT().CleanupNS(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil),
@@ -1532,6 +1229,9 @@ func TestCheckTearDownPauseContainer(t *testing.T) {
 			ContainerJSONBase: &types.ContainerJSONBase{
 				ID:    containerID,
 				State: &types.ContainerState{Pid: containerPid},
+				HostConfig: &dockercontainer.HostConfig{
+					NetworkMode: containerNetworkMode,
+				},
 			},
 		}, nil).MaxTimes(1),
 		mockCNIClient.EXPECT().CleanupNS(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil).MaxTimes(1),
@@ -2440,6 +2140,9 @@ func TestContainerMetadataUpdatedOnRestart(t *testing.T) {
 						State: &types.ContainerState{
 							Health: &types.Health{},
 						},
+						HostConfig: &dockercontainer.HostConfig{
+							NetworkMode: containerNetworkMode,
+						},
 					},
 					Config: &dockercontainer.Config{
 						Labels: labels,
@@ -3262,6 +2965,9 @@ func TestStartFirelensContainerRetryForContainerIP(t *testing.T) {
 		ContainerJSONBase: &types.ContainerJSONBase{
 			ID:    containerID,
 			State: &types.ContainerState{Pid: containerPid},
+			HostConfig: &dockercontainer.HostConfig{
+				NetworkMode: containerNetworkMode,
+			},
 		},
 	}
 
@@ -3269,6 +2975,9 @@ func TestStartFirelensContainerRetryForContainerIP(t *testing.T) {
 		ContainerJSONBase: &types.ContainerJSONBase{
 			ID:    containerID,
 			State: &types.ContainerState{Pid: containerPid},
+			HostConfig: &dockercontainer.HostConfig{
+				NetworkMode: containerNetworkMode,
+			},
 		},
 		NetworkSettings: &types.NetworkSettings{
 			DefaultNetworkSettings: types.DefaultNetworkSettings{
